+++ frontend/src/App.jsx
@@ -17,6 +17,7 @@ import Settings from './pages/Settings.jsx'
 import Account from './pages/Account.jsx'
 import AccessSheet from './pages/AccessSheet.jsx'
 import SystemControl from './pages/SystemControl.jsx'
+import Approvals from './pages/Approvals.jsx'
 
 function PrivateRoute({ children, roles, caps, bare }) {
   const { user } = useAuth()
@@ -45,6 +46,7 @@ export default function App() {
       <Route path="/articles/:id" element={<PrivateRoute><ArticleDetail /></PrivateRoute>} />
       <Route path="/offen" element={<PrivateRoute><OpenIssues /></PrivateRoute>} />
       <Route path="/personen" element={<PrivateRoute caps={['persons']}><Persons /></PrivateRoute>} />
+      <Route path="/genehmigungen" element={<PrivateRoute caps={['articles']}><Approvals /></PrivateRoute>} />
       <Route path="/import" element={<PrivateRoute caps={['export']}><ImportPage /></PrivateRoute>} />
       <Route path="/meine-artikel" element={<PrivateRoute><MyArticles /></PrivateRoute>} />
       <Route path="/scan" element={<PrivateRoute caps={['issues']}><MaterialScan /></PrivateRoute>} />